cleaned up invitations component, moved handle-click functions to separate file
